requirements.txt
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
faiss-cpu==1.7.4
filelock==3.18.0
Flask==3.0.0
fsspec==2025.5.1
huggingface-hub==0.16.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
nltk==3.9.1
numpy==1.24.3
packaging==25.0
pillow==11.2.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.31.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==2.2.2
sentencepiece==0.2.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.13.3
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
transformers==4.33.2
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
